Prelude functions are clock polymorphic
Before, the prelude functions operated on type "Signal", and the
explicit prelude functions operated on type "Signal' clk", and
had an extra "SClock clk" argument.

Now, the prelude functions operate on type "Signal' clk", and so
do the explicit prelude functions. What remains the same is that
the explicit prelude functions still have an extra "SClock clk"
argument.

This was made possible by the injectivity annotation for the
Unbundled' type family, which ensures that Bundle's members
no longer need an "SClock clk" argument.